fix: rename pages directory to views to resolve Next.js build error
