+++ src/app/(auth)/login/driver/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from '@/pages/(auth)/login/LoginPage';
+import LoginPage from '@/views/(auth)/login/LoginPage';
 
 export default function DriverLoginPage() {
   return <LoginPage usertype="DRIVER" />;
