+++ src/app/(auth)/login/user/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from '@/pages/(auth)/login/LoginPage';
+import LoginPage from '@/views/(auth)/login/LoginPage';
 
 export default function UserLoginPage() {
   return <LoginPage usertype="USER" />;
